requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
streamlit==1.36.0
google-genai==0.3.0
pydantic==2.8.2
python-multipart==0.0.9
requests==2.32.3
pandas>=2.1.0,<2.3
altair==5.3.0
pillow>=10.0
python-dotenv==1.0.1
PyJWT
bcrypt
pytest
